Add files via upload
Based on research on YouTube, I discovered a way to improve wutering waves' performance by messing with the game files. 
The "Engine.ini" file has some Unreal Engine variables that are capable of changing the graphics, resolution and effects of 
the game (in fact it does not have variables within this file, you must add them based on your knowledge). I left this file 
with the best options to gain more fps, you can change and add others if you want.